emu: Some frontend QOC improvements
Both:
- mmc id now also scan from folder name. like vanessa suggested here, the mmc id in config.yml is fallback. You will be notified if the emulator founds mmc in folder name

Android:
- Add quick switch devices button
- Add background to config profile. Can choose to keep/not keep aspect ratio
- Clean up the menu and separate n-gage and n-gage 2.0 to a separate group

PC:
- Add discord rich presence, can be disable in settings. Dev build disable by default
- Emulator now restart back to app list when game exit. If game exits, a toast notification will say the reason the game exit
- Moved settings to menu bar, some people don't go into File submenu to look for it

+++ src/emu/android/app/src/main/java/com/github/eka2l1/util/FileUtils.java
@@ -20,8 +20,13 @@
 package com.github.eka2l1.util;
 
 import android.content.Context;
+import android.database.Cursor;
+import android.icu.util.Output;
+import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;
+import android.os.ParcelFileDescriptor;
+import android.provider.OpenableColumns;
 import android.util.Log;
 
 import androidx.activity.result.contract.ActivityResultContract;
@@ -29,6 +34,7 @@ import androidx.activity.result.contract.ActivityResultContract;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FilenameFilter;
 import java.io.IOException;
@@ -37,6 +43,8 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.nio.channels.FileChannel;
+import java.nio.file.Files;
+import java.nio.file.StandardCopyOption;
 
 public class FileUtils {
 
@@ -65,7 +73,48 @@ public class FileUtils {
         }
     }
 
+    public static String getFilenameFromURI(Context context, String path) {
+        Cursor returnCursor = context.getContentResolver().query(Uri.parse(path), null, null, null, null);
+        assert returnCursor != null;
+        int nameIndex = returnCursor.getColumnIndex(OpenableColumns.DISPLAY_NAME);
+        returnCursor.moveToFirst();
+        String name = returnCursor.getString(nameIndex);
+        returnCursor.close();
+        return name;
+    }
+
+    public static String getFilenameFromPath(Context context, String path) {
+        String folderName;
+
+        if (path.startsWith("/")) {
+            folderName = (new File(path)).getName();
+        } else {
+            folderName = getFilenameFromURI(context, path);
+        }
+
+        return folderName;
+    }
+
+    public static void copyFileFromURI(Context context, String pointedURI, File dest) throws IOException {
+        Uri uri = Uri.parse(pointedURI);
+        copyFileFromInputStream(context.getContentResolver().openInputStream(uri), dest);
+    }
+
+    public static void copyFileFromInputStream(InputStream source, File dest) throws IOException {
+        byte[] buffer = new byte[2048];
+        int read = 0;
+
+        OutputStream stream = new FileOutputStream(dest);
+
+        while ((read = source.read(buffer)) != -1) {
+            stream.write(buffer, 0, read);
+        }
+    }
+
     public static void copyFileUsingChannel(File source, File dest) throws IOException {
+        if (!source.exists()) {
+            return;
+        }
         try (FileChannel sourceChannel = new FileInputStream(source).getChannel();
              FileChannel destChannel = new FileOutputStream(dest).getChannel()) {
             destChannel.transferFrom(sourceChannel, 0, sourceChannel.size());
